Use only one OS socket per McastRxSocket on Unix IPv4 (#21)
Until now, I had thought it was impossible, on Unix OSs, to subscribe to
multiple multicast groups on one socket. And while I still suspect that
this is the case for IPv6, thanks to my colleagues Jasper and Pavel, we
have discovered that this is not true for IPv4. By setting the
IP_MULTICAST_ALL option to false, the socket will only receive from the
exact multicast addresses it's subscribed to, rather than all multicasts
arriving to that interface + port combo.

I added a new test to confirm this behavior, and verified that the tests
(specifically test_v4_loopback_multiple) failed on Linux until I added
the IP_MULTICAST_ALL option.

I also updated the README and changelog to reflect the changes.