Implement AsyncMcastRxSocket, providing asynchronous I/O support (#22)
This PR adds AsyncMcastRxSocket, which is a variant of McastRxSocket
whose recv() and recvfrom() options are awaitable coroutines. This
should make it way more convenient to use multicast from within asyncio
event-based code!

Implementing this was actually a bit difficult, because Python asyncio
is not very well set up to receive from multiple sockets at the same
time. At first I tried [this SO
post](https://stackoverflow.com/a/48250808/7083698)'s recommendation of
using `EventLoop.add_reader()`, but soon found out the hard way that
this is not available on Windows. Then I looked into a method that my
colleague Colin H used in an internal tool. His solution works by
creating a bunch of Tasks (one for each socket) that call

```python
await asyncio.wait_for(asyncio.get_event_loop().sock_recv(sock, 4096), timeout)
```

and then calling `asyncio.gather()` on those tasks. This solution works,
but with a big caveat: it requires all of the tasks to either complete
or be canceled before gather() will return. So, if you call recv() with
a timeout of 1 second, and one socket has a packet right away but the
other socket has nothing, it will wait the full 1 second before
returning.

The way to avoid this is to use
`asyncio.wait(return_when=asyncio.FIRST_COMPLETED)`, which will return
when any task has completed. That works, but then you're stuck with a
bunch of other tasks that have not yet completed but still represent
active recv() calls on the socket, and canceling those seems a bit
sketchy (what if they completed and have packets to return?).

So, I set up the class to create and start a Task for each socket in
`__enter__`. Then, in recvfrom(), we find one task that has completed,
grab the resultant packet, and then recreate the task to receive the
next packet. This seems to work, and is actually quite a bit simpler
than what I started with!

Note: The new additions currently require Python 3.11 or newer, and will
not be available for import on older python versions. This is because
`EventLoop.sock_recvfrom()` was added in 3.11. `sock_recv()` exists in
3.9 and 3.10, but using this method would prevent you from accessing the
source address of the multicast, which is kinda limiting. So, I felt
like a reasonable medium was to hide the new additions behind a Python
version check.

Note 2: the only thing I think is a bit sketchy about this
implementation is that what happens when it's closed. We try and cancel
all the recvfrom tasks, but it isn't actually documented what happens
when the result of `EventLoop.sock_recvfrom()` is canceled. What if it's
received a packet? I feel like there's a good chance that packet can
just be dropped. So, closing and reopening the AsyncMcastRxSocket could
result in dropping packets that were queued in the OS. However, closing
a socket already means you kinda don't care about what is being received
on it, so I *think* this should be OK.